econ-ml-portfolio/requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
pytorch-lightning>=2.2.0
einops>=0.7.0

# ── Transformers / NLP ────────────────────────────────────────────────────────
transformers>=4.40.0
datasets>=2.19.0
sentence-transformers>=2.7.0
tokenizers>=0.19.0

# ── Graph Neural Networks ─────────────────────────────────────────────────────
# torch-geometric installed separately in Dockerfile (needs matching torch version)
networkx>=3.3

# ── Time Series ───────────────────────────────────────────────────────────────
pytorch-forecasting>=1.0.0

# ── Data science ─────────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
scipy>=1.13.0
statsmodels>=0.14.0

# ── Geospatial / Satellite ────────────────────────────────────────────────────
rasterio>=1.3.10
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
pyogrio>=0.7.0

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.21.0
umap-learn>=0.5.6

# ── Explainability ────────────────────────────────────────────────────────────
shap>=0.45.0
captum>=0.7.0

# ── Data retrieval ────────────────────────────────────────────────────────────
wbgapi>=1.0.12        # World Bank API
pytrends>=4.9.2       # Google Trends
requests>=2.31.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# ── Config / utilities ────────────────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.12.0

# ── Dev / testing ─────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-cov>=5.0.0
black>=24.0.0
ruff>=0.4.0
